BAEL-2388: moved examples java-collections-maps -> core-java-long. Fixing String comparisons
