+++ core-java-lang/src/main/java/com/baeldung/equalshashcode/Team.java
@@ -1,4 +1,4 @@
-package com.baeldung.map.hashcode;
+package com.baeldung.equalshashcode;
 
 class Team {
