+++ core-java-lang/src/test/java/com/baeldung/equalshashcode/MoneyUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.map.hashcode;
+package com.baeldung.equalshashcode;
 
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.assertFalse;
