+++ core-java-lang/src/test/java/com/baeldung/equalshashcode/TeamUnitTest.java
@@ -1,4 +1,4 @@
-package com.baeldung.map.hashcode;
+package com.baeldung.equalshashcode;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
