--- java-collections-maps/src/main/java/com/baeldung/map/hashCode/equalsverifier_notes.txt
@@ -1,1 +1,0 @@
- Non-nullity: hashCode throws NullPointerException on field city.
